+++ src/base.js
@@ -1,3 +1,6 @@
+import Label from './components/label'
+import Site from './components/site'
+
 export default class Base {
 	/**
 	 * To be overwritten by child classes.
@@ -99,7 +102,20 @@ export default class Base {
 	async init() {
 		this.$watch(this.filterKeys.join(', '), () => this.filterModels())
 
-		let response = await CS.fetchJSON(this.resourceModule, this.options)
+		let response = await CS.fetchJSON(this.resourceModule, Object.assign({}, this.options))
+
+		// set the default image to the brand emblem
+		if (response.hasOwnProperty('brand')) {
+			this.emblemImage = response.brand.emblem[512].url
+		}
+
+		/** 
+		 * For efficiency, the Organisation response sends over the labels once
+		 * on page 1, rather than on every Organisation.
+		 */
+		if (response.hasOwnProperty('labels')) {
+			response.labels.forEach(label => this.labels.push(new Label(label)))
+		}
 
 		if (response.hasOwnProperty('configuration')) {
 			// new style configuration data
@@ -111,15 +127,45 @@ export default class Base {
 			})
 		} else {
 			// old style flat array
-			window.test = this
 			response.forEach(model => {
 				this.modelsAll.push(this.buildModelObject(model))
 			})
 		}
 
+		/** 
+		 * For efficiency, the Organisation response sends over the sites once
+		 * on page 1, rather than on every Organisation.
+		 */
+		if (response.hasOwnProperty('sites')) {
+			response.sites.forEach(site => this.sites.push(new Site(site)))
+		}
+
 		this.postInit()
 
 		this.$nextTick(() => this.filterModels())
+
+		// go and fetch the rest of the paginated data
+		if (response.hasOwnProperty('pagination')) {
+			// don't do anything if we already have all the data
+			if (response.pagination.num_results <= response.pagination.results_per_page) return
+
+			
+				this.$nextTick(() => {
+					let promises = []
+					for (let page = 2; page <= response.pagination.totalPages; page++) {
+						let options = Object.assign({}, this.options)
+						options.page = page
+						promises.push(CS.fetchJSON(this.resourceModule, options)
+							.then(response => response.data.forEach(model => {
+								this.modelsAll.push(this.buildModelObject(model))
+							})))
+					}
+
+					Promise.allSettled(promises).then(() => this.filterModels())
+				})
+			
+		}
+		
 	}
 
 	/**